+++ lib/basic/indexed_file.dart
@@ -12,6 +12,12 @@ class IndexedFile extends Widget {
         ? IndexedFile._indexed[name]
         : 0;
   }
+  // get the id of a name and increase usage.
+  static int getIndexedAndIncrease(String name) {
+    final id = getIndexed(name);
+    IndexedFile._indexed[name] = id + 1;
+    return id;
+  }
 
   final String name;
   final String custom;
@@ -22,8 +28,14 @@ class IndexedFile extends Widget {
 
   /// The IndexedFile behaves similar to File. Additionally it makes sure that each File, created with IndexedFile, is unique and does not get overwritten.
   /// In order to do that IndexedFile saves for each inputted name an id, which gets incremented after each use.
-  IndexedFile(this.name,
-      {this.child, this.custom, this.path, this.execute = false, this.pack});
+  IndexedFile(
+    this.name, {
+    this.child,
+    this.custom,
+    this.path,
+    this.execute = false,
+    this.pack,
+  });
 
   // gets the File id.
   int getId() {